requirements.txt
pygame==2.0.1
requests==2.26.0
alive-progress==2.0.0
load_dotenv==20.2.3